+++ 2. Variables and interactive programs/3.py
@@ -1,5 +1,5 @@
-length = int(input("Enter the length of the rectangle: "))
-width = int(input("Enter the width of the rectangle: "))
+length = float(input("Enter the length of the rectangle: "))
+width = float(input("Enter the width of the rectangle: "))
 
 perimeter = 2 * length + 2 * width
 area = length * width
